refactor: rename Message and pull out into own file
