+++ PrismMessenger/User/User.swift
@@ -50,46 +50,3 @@ final class User: Identifiable {
         return publicPrekeys
     }
 }
-
-
-@Model
-final class MessageData: Identifiable {
-    @Attribute(.unique) var id: UUID
-    var content: String
-    var timestamp: Date
-    var isFromMe: Bool
-    var status: MessageStatus
-    
-    // Server-related properties
-    var serverId: UUID?
-    var serverTimestamp: Date?
-    
-    // Reference back to parent chat
-    var chat: Chat?
-    
-    init(content: String, isFromMe: Bool, status: MessageStatus = .sent) {
-        self.id = UUID()
-        self.content = content
-        self.timestamp = Date()
-        self.isFromMe = isFromMe
-        self.status = status
-    }
-}
-
-enum MessageStatus: String, Codable {
-    case sending
-    case sent
-    case delivered
-    case read
-    case failed
-}
-
-extension ModelContext {
-    var sqliteCommand: String {
-        if let url = container.configurations.first?.url.path(percentEncoded: false) {
-            "sqlite3 \"\(url)\""
-        } else {
-            "No SQLite database found."
-        }
-    }
-}
